@@ -6,7 +6,7 @@ require 'rack/test'
 
 describe MessageBus::Rack::Middleware do
   include Rack::Test::Methods
-  let(:extra_middleware){nil}
+  let(:extra_middleware) { nil }
 
   before do
     bus = @bus = MessageBus::Instance.new
@@ -15,10 +15,10 @@ describe MessageBus::Rack::Middleware do
 
     e_m = extra_middleware
     builder = Rack::Builder.new {
-      use FakeAsyncMiddleware, :message_bus => bus
+      use FakeAsyncMiddleware, message_bus: bus
       use e_m if e_m
-      use MessageBus::Rack::Middleware, :message_bus => bus
-      run lambda {|env| [500, {'Content-Type' => 'text/html'}, 'should not be called' ]}
+      use MessageBus::Rack::Middleware, message_bus: bus
+      run lambda { |_env| [500, { 'Content-Type' => 'text/html' }, 'should not be called'] }
     }
 
     @async_middleware = builder.to_app
@@ -62,12 +62,12 @@ describe MessageBus::Rack::Middleware do
       middleware = @async_middleware
       bus = @bus
 
-      @bus.extra_response_headers_lookup do |env|
-        {"FOO" => "BAR"}
+      @bus.extra_response_headers_lookup do |_env|
+        { "FOO" => "BAR" }
       end
 
       Thread.new do
-        wait_for(2000) {middleware.in_async?}
+        wait_for(2000) { middleware.in_async? }
         bus.publish "/foo", "םוֹלשָׁ"
       end
 
@@ -111,26 +111,45 @@ describe MessageBus::Rack::Middleware do
     include LongPolling
   end
 
-  describe "diagnostics" do
+  describe "start listener" do
+    let(:app) { ->(_) { [200, {}, []] } }
 
+    it "never subscribes" do
+      bus = MessageBus::Instance.new
+      bus.off
+
+      middleware = MessageBus::Rack::Middleware.new(app, message_bus: bus)
+
+      middleware.started_listener.must_equal false
+    end
+  end
+
+  describe "diagnostics" do
     it "should return a 403 if a user attempts to get at the _diagnostics path" do
       get "/message-bus/_diagnostics"
       last_response.status.must_equal 403
     end
 
     it "should get a 200 with html for an authorized user" do
-      def @bus.is_admin_lookup; proc{|_| true} end
+
+      def @bus.is_admin_lookup
+        proc { |_| true }
+      end
+
       get "/message-bus/_diagnostics"
       last_response.status.must_equal 200
     end
 
     it "should get the script it asks for" do
-      def @bus.is_admin_lookup; proc{|_| true} end
+
+      def @bus.is_admin_lookup
+        proc { |_| true }
+      end
+
       get "/message-bus/_diagnostics/assets/message-bus.js"
       last_response.status.must_equal 200
-      last_response.content_type.must_equal "text/javascript;"
+      last_response.content_type.must_equal "application/javascript;charset=UTF-8"
     end
-
   end
 
   describe "polling" do
@@ -139,17 +158,16 @@ describe MessageBus::Rack::Middleware do
     end
 
     it "should include access control headers" do
-      @bus.extra_response_headers_lookup do |env|
-        {"FOO" => "BAR"}
+      @bus.extra_response_headers_lookup do |_env|
+        { "FOO" => "BAR" }
       end
 
       client_id = "ABCD"
 
       # client always keeps a list of channels with last message id they got on each
-      post "/message-bus/#{client_id}", {
-        '/foo' => nil,
-        '/bar' => nil
-      }
+      post "/message-bus/#{client_id}",
+           '/foo' => nil,
+           '/bar' => nil
 
       last_response.headers["FOO"].must_equal "BAR"
     end
@@ -158,39 +176,65 @@ describe MessageBus::Rack::Middleware do
       client_id = "ABCD"
 
       # client always keeps a list of channels with last message id they got on each
-      post "/message-bus/#{client_id}", {
-        '/foo' => nil,
-        '/bar' => nil
-      }
+      post "/message-bus/#{client_id}",
+           '/foo' => nil,
+           '/bar' => nil
+
       last_response.ok?.must_equal true
+    end
+
+    # this means we recover from redis reset
+    it "should understand that larger than position is the same as -1" do
+      @bus.publish('/foo', 'bar')
+      @bus.publish('/baz', 'test')
+      @bus.publish('/boom', 'bang')
+
+      post "/message-bus/ABCD",
+           '/foo' => 1_000_000,
+           '/baz' => @bus.last_id('/baz') + 1,
+           '/boom' => 1_000_000
+
+      last_response.ok?.must_equal true
+      parsed = JSON.parse(last_response.body)
+
+      parsed.length.must_equal 1
+      parsed[0]["channel"].must_equal "/__status"
+      parsed[0]["data"]["/foo"].must_equal @bus.last_id("/foo")
+      parsed[0]["data"]["/boom"].must_equal @bus.last_id("/boom")
     end
 
     it "should correctly understand that -1 means stuff from now onwards" do
+      # even if allow chunked
+      @bus.chunked_encoding_enabled = true
 
-      @bus.publish('foo', 'bar')
+      @bus.publish('/foo', 'bar')
+      @bus.publish('/baz', 'test')
+      @bus.publish('/boom', 'bang')
 
-      post "/message-bus/ABCD", {
-        '/foo' => -1
-      }
+      post "/message-bus/ABCD",
+           '/foo' => -1,
+           '/baz' => @bus.last_id('/baz') + 1,
+           '/boom' => -1
+
       last_response.ok?.must_equal true
       parsed = JSON.parse(last_response.body)
+
       parsed.length.must_equal 1
       parsed[0]["channel"].must_equal "/__status"
-      parsed[0]["data"]["/foo"].must_equal@bus.last_id("/foo")
-
+      parsed[0]["data"]["/foo"].must_equal @bus.last_id("/foo")
+      parsed[0]["data"]["/boom"].must_equal @bus.last_id("/boom")
     end
 
     it "should respond with the data if messages exist in the backlog" do
-      id =@bus.last_id('/foo')
+      id = @bus.last_id('/foo')
 
       @bus.publish("/foo", "barbs")
       @bus.publish("/foo", "borbs")
 
       client_id = "ABCD"
-      post "/message-bus/#{client_id}", {
-        '/foo' => id,
-        '/bar' => nil
-      }
+      post "/message-bus/#{client_id}",
+           '/foo' => id,
+           '/bar' => nil
 
       parsed = JSON.parse(last_response.body)
       parsed.length.must_equal 2
@@ -199,37 +243,32 @@ describe MessageBus::Rack::Middleware do
     end
 
     it "should have no cross talk" do
-
       seq = 0
       @bus.site_id_lookup do
-        (seq+=1).to_s
+        (seq += 1).to_s
       end
 
       # published on channel 1
       msg = @bus.publish("/foo", "test")
 
       # subscribed on channel 2
-      post "/message-bus/ABCD", {
-        '/foo' => (msg-1)
-      }
+      post "/message-bus/ABCD",
+           '/foo' => (msg - 1)
 
       parsed = JSON.parse(last_response.body)
       parsed.length.must_equal 0
-
     end
 
     it "should have global cross talk" do
-
       seq = 0
       @bus.site_id_lookup do
-        (seq+=1).to_s
+        (seq += 1).to_s
       end
 
       msg = @bus.publish("/global/foo", "test")
 
-      post "/message-bus/ABCD", {
-        '/global/foo' => (msg-1)
-      }
+      post "/message-bus/ABCD",
+           '/global/foo' => (msg - 1)
 
       parsed = JSON.parse(last_response.body)
       parsed.length.must_equal 1
@@ -237,64 +276,67 @@ describe MessageBus::Rack::Middleware do
 
     it "should not get consumed messages" do
       @bus.publish("/foo", "barbs")
-      id =@bus.last_id('/foo')
+      id = @bus.last_id('/foo')
 
       client_id = "ABCD"
-      post "/message-bus/#{client_id}", {
-        '/foo' => id
-      }
+      post "/message-bus/#{client_id}",
+           '/foo' => id
 
       parsed = JSON.parse(last_response.body)
       parsed.length.must_equal 0
     end
 
     it "should filter by user correctly" do
-      id =@bus.publish("/foo", "test", user_ids: [1])
-      @bus.user_id_lookup do |env|
+      id = @bus.publish("/foo", "test", user_ids: [1])
+      @bus.user_id_lookup do |_env|
         0
       end
 
       client_id = "ABCD"
-      post "/message-bus/#{client_id}", {
-        '/foo' => id - 1
-      }
+      post "/message-bus/#{client_id}",
+           '/foo' => id - 1
 
       parsed = JSON.parse(last_response.body)
-      parsed.length.must_equal 0
+      parsed.length.must_equal 1
 
-      @bus.user_id_lookup do |env|
+      message = parsed.first
+
+      message["channel"].must_equal "/__status"
+      message["data"].must_equal("/foo" => 1)
+
+      @bus.user_id_lookup do |_env|
         1
       end
 
-      post "/message-bus/#{client_id}", {
-        '/foo' => id - 1
-      }
+      post "/message-bus/#{client_id}",
+           '/foo' => id - 1
 
       parsed = JSON.parse(last_response.body)
       parsed.length.must_equal 1
     end
 
     it "should filter by group correctly" do
-      id =@bus.publish("/foo", "test", group_ids: [3,4,5])
-      @bus.group_ids_lookup do |env|
-        [0,1,2]
+      id = @bus.publish("/foo", "test", group_ids: [3, 4, 5])
+      @bus.group_ids_lookup do |_env|
+        [0, 1, 2]
       end
 
       client_id = "ABCD"
-      post "/message-bus/#{client_id}", {
-        '/foo' => id - 1
-      }
+      post "/message-bus/#{client_id}",
+           '/foo' => id - 1
 
       parsed = JSON.parse(last_response.body)
-      parsed.length.must_equal 0
+      message = parsed.first
 
-      @bus.group_ids_lookup do |env|
-        [1,7,4,100]
+      message["channel"].must_equal "/__status"
+      message["data"].must_equal("/foo" => 1)
+
+      @bus.group_ids_lookup do |_env|
+        [1, 7, 4, 100]
       end
 
-      post "/message-bus/#{client_id}", {
-        '/foo' => id - 1
-      }
+      post "/message-bus/#{client_id}",
+           '/foo' => id - 1
 
       parsed = JSON.parse(last_response.body)
       parsed.length.must_equal 1
@@ -302,32 +344,41 @@ describe MessageBus::Rack::Middleware do
 
     it "can decode a JSON encoded request" do
       id = @bus.last_id('/foo')
-      @bus.publish("/foo", {json: true})
-      post( "/message-bus/1234",
-            JSON.generate({'/foo' => id}),
-            { "CONTENT_TYPE" => "application/json" })
-      JSON.parse(last_response.body).first["data"].must_equal({'json' => true})
+      @bus.publish("/foo", json: true)
+      post("/message-bus/1234",
+           JSON.generate('/foo' => id),
+           "CONTENT_TYPE" => "application/json")
+      JSON.parse(last_response.body).first["data"].must_equal('json' => true)
     end
 
     describe "on_middleware_error handling" do
       it "allows error handling of middleware failures" do
-
-        @bus.on_middleware_error do |env, err|
+        @bus.on_middleware_error do |_env, err|
           if ArgumentError === err
-            [407,{},[]]
+            [407, {}, []]
           end
         end
 
-        @bus.group_ids_lookup do |env|
+        @bus.group_ids_lookup do |_env|
           raise ArgumentError
         end
 
-        post( "/message-bus/1234",
-            JSON.generate({'/foo' => 1}),
-            { "CONTENT_TYPE" => "application/json" })
+        post("/message-bus/1234",
+             JSON.generate('/foo' => 1),
+             "CONTENT_TYPE" => "application/json")
 
         last_response.status.must_equal 407
+      end
 
+      it "does not handle exceptions from downstream middleware" do
+        @bus.on_middleware_error do |_env, err|
+          [404, {}, []]
+        end
+
+        get("/")
+
+        last_response.status.must_equal 500
+        last_response.body.must_equal 'should not be called'
       end
     end
 
@@ -341,7 +392,7 @@ describe MessageBus::Rack::Middleware do
           end
 
           def call(env)
-            @app.call(env.merge('message_bus.channels'=>{'/foo'=>0}))
+            @app.call(env.merge('message_bus.channels' => { '/foo' => 0 }))
           end
         end
       end
@@ -351,16 +402,14 @@ describe MessageBus::Rack::Middleware do
         foo_id = @bus.publish("/foo", "testfoo")
         bar_id = @bus.publish("/bar", "testbar")
 
-        post "/message-bus/ABCD", {
-          '/foo' => foo_id - 1
-        }
+        post "/message-bus/ABCD",
+             '/foo' => foo_id - 1
 
         parsed = JSON.parse(last_response.body)
         parsed.first['data'].must_equal 'testfoo'
 
-        post "/message-bus/ABCD", {
-          '/bar' => bar_id - 1
-        }
+        post "/message-bus/ABCD",
+             '/bar' => bar_id - 1
 
         parsed = JSON.parse(last_response.body)
         parsed.first['data'].must_equal 'testfoo'
