@@ -1,43 +1,49 @@
+# frozen_string_literal: true
+
 require 'json'
 
 # our little message bus, accepts long polling and polling
 module MessageBus::Rack; end
 
+# Accepts requests from subscribers, validates and authenticates them,
+# delivers existing messages from the backlog and informs a
+# `MessageBus::ConnectionManager` of a connection which is remaining open.
 class MessageBus::Rack::Middleware
-
-  def start_listener
-    unless @started_listener
-
-      thin = defined?(Thin::Server) && ObjectSpace.each_object(Thin::Server).to_a.first
-      thin_running = thin && thin.running?
-
-      @subscription = @bus.subscribe do |msg|
-        run = proc do
-          begin
-            @connection_manager.notify_clients(msg) if @connection_manager
-          rescue
-            @bus.logger.warn "Failed to notify clients: #{$!} #{$!.backtrace}"
-          end
-        end
-
-        if thin_running
-          EM.next_tick(&run)
-        else
-          MessageBus.timer.queue(&run)
-        end
-
-        @started_listener = true
-      end
-    end
+  # @param [Array<MessageBus::Message>] backlog a list of messages for delivery
+  # @return [JSON] a JSON representation of the backlog, compliant with the
+  #   subscriber API specification
+  def self.backlog_to_json(backlog)
+    m = backlog.map do |msg|
+      {
+        global_id: msg.global_id,
+        message_id: msg.message_id,
+        channel: msg.channel,
+        data: msg.data
+      }
+    end.to_a
+    JSON.dump(m)
   end
 
+  # @return [Boolean] whether the message listener (subscriber) is started or not)
+  attr_reader :started_listener
+
+  # Sets up the middleware to receive subscriber client requests and begins
+  # listening for messages published on the bus for re-distribution (unless
+  # the bus is disabled).
+  #
+  # @param [Proc] app the rack app
+  # @param [Hash] config
+  # @option config [MessageBus::Instance] :message_bus (`MessageBus`) a specific instance of message_bus
   def initialize(app, config = {})
     @app = app
     @bus = config[:message_bus] || MessageBus
     @connection_manager = MessageBus::ConnectionManager.new(@bus)
-    self.start_listener
+    @started_listener = false
+    start_listener unless @bus.off?
   end
 
+  # Stops listening for messages on the bus
+  # @return [void]
   def stop_listener
     if @subscription
       @bus.unsubscribe(&@subscription)
@@ -45,30 +51,25 @@ class MessageBus::Rack::Middleware
     end
   end
 
-  def self.backlog_to_json(backlog)
-    m = backlog.map do |msg|
-      {
-        :global_id => msg.global_id,
-        :message_id => msg.message_id,
-        :channel => msg.channel,
-        :data => msg.data
-      }
-    end.to_a
-    JSON.dump(m)
-  end
-
+  # Process an HTTP request from a subscriber client
+  # @param [Rack::Request::Env] env the request environment
   def call(env)
-
     return @app.call(env) unless env['PATH_INFO'] =~ /^\/message-bus\//
 
+    handle_request(env)
+  end
+
+  private
+
+  def handle_request(env)
     # special debug/test route
-    if @bus.allow_broadcast? && env['PATH_INFO'] == '/message-bus/broadcast'.freeze
-        parsed = Rack::Request.new(env)
-        @bus.publish parsed["channel".freeze], parsed["data".freeze]
-        return [200,{"Content-Type".freeze => "text/html".freeze},["sent"]]
+    if @bus.allow_broadcast? && env['PATH_INFO'] == '/message-bus/broadcast'
+      parsed = Rack::Request.new(env)
+      @bus.publish parsed["channel"], parsed["data"]
+      return [200, { "Content-Type" => "text/html" }, ["sent"]]
     end
 
-    if env['PATH_INFO'].start_with? '/message-bus/_diagnostics'.freeze
+    if env['PATH_INFO'].start_with? '/message-bus/_diagnostics'
       diags = MessageBus::Rack::Diagnostics.new(@app, message_bus: @bus)
       return diags.call(env)
     end
@@ -97,8 +98,8 @@ class MessageBus::Rack::Middleware
       request = Rack::Request.new(env)
       is_json = request.content_type && request.content_type.include?('application/json')
       data = is_json ? JSON.parse(request.body.read) : request.POST
-      data.each do |k,v|
-        if k == "__seq".freeze
+      data.each do |k, v|
+        if k == "__seq"
           client.seq = v.to_i
         else
           client.subscribe(k, v)
@@ -113,7 +114,7 @@ class MessageBus::Rack::Middleware
     headers["Expires"] = "0"
 
     if @bus.extra_response_headers_lookup
-      @bus.extra_response_headers_lookup.call(env).each do |k,v|
+      @bus.extra_response_headers_lookup.call(env).each do |k, v|
         headers[k] = v
       end
     end
@@ -123,11 +124,11 @@ class MessageBus::Rack::Middleware
     end
 
     long_polling = @bus.long_polling_enabled? &&
-                   env['QUERY_STRING'] !~ /dlp=t/.freeze &&
+                   env['QUERY_STRING'] !~ /dlp=t/ &&
                    @connection_manager.client_count < @bus.max_active_clients
 
-    allow_chunked = env['HTTP_VERSION'.freeze] == 'HTTP/1.1'.freeze
-    allow_chunked &&= !env['HTTP_DONT_CHUNK'.freeze]
+    allow_chunked = env['HTTP_VERSION'] == 'HTTP/1.1'
+    allow_chunked &&= !env['HTTP_DONT_CHUNK']
     allow_chunked &&= @bus.chunked_encoding_enabled?
 
     client.use_chunked = allow_chunked
@@ -135,9 +136,9 @@ class MessageBus::Rack::Middleware
     backlog = client.backlog
 
     if backlog.length > 0 && !allow_chunked
-      client.cancel
+      client.close
       @bus.logger.debug "Delivering backlog #{backlog} to client #{client_id} for user #{user_id}"
-      [200, headers, [self.class.backlog_to_json(backlog)] ]
+      [200, headers, [self.class.backlog_to_json(backlog)]]
     elsif long_polling && env['rack.hijack'] && @bus.rack_hijack_enabled?
       io = env['rack.hijack'].call
       # TODO disable client till deliver backlog is called
@@ -159,7 +160,7 @@ class MessageBus::Rack::Middleware
         response = Thin::AsyncResponse.new(env)
       end
 
-      headers.each do |k,v|
+      headers.each do |k, v|
         response.headers[k] = v
       end
 
@@ -179,11 +180,10 @@ class MessageBus::Rack::Middleware
 
       throw :async
     else
-      [200, headers, ["[]"]]
+      [200, headers, [self.class.backlog_to_json(backlog)]]
     end
-
   rescue => e
-    if @bus.on_middleware_error && result=@bus.on_middleware_error.call(env, e)
+    if @bus.on_middleware_error && result = @bus.on_middleware_error.call(env, e)
       result
     else
       raise
@@ -203,14 +203,39 @@ class MessageBus::Rack::Middleware
   def add_client_with_timeout(client)
     @connection_manager.add_client(client)
 
-    client.cleanup_timer = MessageBus.timer.queue( @bus.long_polling_interval.to_f / 1000) {
+    client.cleanup_timer = @bus.timer.queue(@bus.long_polling_interval.to_f / 1000) {
       begin
-        client.cleanup_timer = nil
-        client.ensure_closed!
+        client.close
         @connection_manager.remove_client(client)
       rescue
         @bus.logger.warn "Failed to clean up client properly: #{$!} #{$!.backtrace}"
       end
     }
+  end
+
+  def start_listener
+    unless @started_listener
+
+      thin = defined?(Thin::Server) && ObjectSpace.each_object(Thin::Server).to_a.first
+      thin_running = thin && thin.running?
+
+      @subscription = @bus.subscribe do |msg|
+        run = proc do
+          begin
+            @connection_manager.notify_clients(msg) if @connection_manager
+          rescue
+            @bus.logger.warn "Failed to notify clients: #{$!} #{$!.backtrace}"
+          end
+        end
+
+        if thin_running
+          EM.next_tick(&run)
+        else
+          @bus.timer.queue(&run)
+        end
+
+        @started_listener = true
+      end
+    end
   end
 end
