@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module MessageBus
-  VERSION = "2.0.2"
+  VERSION = "2.2.1"
 end
