@@ -2,13 +2,15 @@ require_relative '../../spec_helper'
 require 'message_bus'
 
 describe MessageBus::Client do
-
   describe "subscriptions" do
+    def setup_client(client_id)
+      MessageBus::Client.new client_id: client_id, message_bus: @bus
+    end
 
     before do
       @bus = MessageBus::Instance.new
       @bus.configure(MESSAGE_BUS_CONFIG)
-      @client = MessageBus::Client.new client_id: 'abc', message_bus: @bus
+      @client = setup_client('abc')
     end
 
     after do
@@ -25,17 +27,18 @@ describe MessageBus::Client do
 
       while line = lines.shift
         break if line == ""
-        name,val = line.split(": ")
+
+        name, val = line.split(": ")
         headers[name] = val
       end
 
-      length = nil
       while line = lines.shift
         length = line.to_i(16)
         break if length == 0
+
         rest = lines.join("\r\n")
         chunks << rest[0...length]
-        lines = (rest[length+2..-1] || "").split("\r\n")
+        lines = (rest[length + 2..-1] || "").split("\r\n")
       end
 
       # split/join gets tricky
@@ -45,15 +48,15 @@ describe MessageBus::Client do
     end
 
     def parse_chunk(data)
-      payload,_ = data.split(/\r\n\|\r\n/m)
+      payload, _ = data.split(/\r\n\|\r\n/m)
       JSON.parse(payload)
     end
 
     it "can chunk replies" do
       @client.use_chunked = true
-      r,w = IO.pipe
+      r, w = IO.pipe
       @client.io = w
-      @client.headers = {"Content-Type" => "application/json; charset=utf-8"}
+      @client.headers = { "Content-Type" => "application/json; charset=utf-8" }
       @client << MessageBus::Message.new(1, 1, '/test', 'test')
       @client << MessageBus::Message.new(2, 2, '/test', "a|\r\n|\r\n|b")
 
@@ -80,7 +83,7 @@ describe MessageBus::Client do
 
       data[-5..-1].must_equal "0\r\n\r\n"
 
-      _,_,chunks = http_parse("HTTP/1.1 200 OK\r\n\r\n" << data)
+      _, _, chunks = http_parse("HTTP/1.1 200 OK\r\n\r\n" << data)
 
       chunks.length.must_equal 2
 
@@ -91,7 +94,6 @@ describe MessageBus::Client do
       # end with []
       chunk2 = parse_chunk(chunks[1])
       chunk2.length.must_equal 0
-
     end
 
     it "does not bleed data accross sites" do
@@ -112,11 +114,43 @@ describe MessageBus::Client do
       log[0].data.must_equal("/hello" => 0)
     end
 
+    it "allows negative subscribes to look behind" do
+      @bus.publish '/hello', 'world'
+      @bus.publish '/hello', 'sam'
+
+      @client.subscribe('/hello', -2)
+
+      log = @client.backlog
+      log.length.must_equal(1)
+      log[0].data.must_equal("sam")
+    end
+
     it "provides status" do
       @client.subscribe('/hello', -1)
       log = @client.backlog
       log.length.must_equal 1
       log[0].data.must_equal("/hello" => 0)
+    end
+
+    it 'provides status updates to clients that are not allowed to a message' do
+      another_client = setup_client('def')
+      clients = [@client, another_client]
+
+      channel = SecureRandom.hex
+
+      clients.each { |client| client.subscribe(channel, nil) }
+
+      @bus.publish(channel, "world", client_ids: ['abc'])
+
+      log = @client.backlog
+      log.length.must_equal 1
+      log[0].channel.must_equal channel
+      log[0].data.must_equal 'world'
+
+      log = another_client.backlog
+      log.length.must_equal 1
+      log[0].channel.must_equal '/__status'
+      log[0].data.must_equal(channel => 1)
     end
 
     it "should provide a list of subscriptions" do
@@ -135,7 +169,7 @@ describe MessageBus::Client do
 
     it "allows only client_id in list if message contains client_ids" do
       @message = MessageBus::Message.new(1, 2, '/test', 'hello')
-      @message.client_ids = ["1","2"]
+      @message.client_ids = ["1", "2"]
       @client.client_id = "2"
       @client.allowed?(@message).must_equal true
 
@@ -145,27 +179,25 @@ describe MessageBus::Client do
 
     describe "targetted at group" do
       before do
-        @message = MessageBus::Message.new(1,2,'/test', 'hello')
-        @message.group_ids = [1,2,3]
+        @message = MessageBus::Message.new(1, 2, '/test', 'hello')
+        @message.group_ids = [1, 2, 3]
       end
 
       it "denies users that are not members of group" do
-        @client.group_ids = [77,0,10]
+        @client.group_ids = [77, 0, 10]
         @client.allowed?(@message).must_equal false
       end
 
       it "allows users that are members of group" do
-        @client.group_ids = [1,2,3]
+        @client.group_ids = [1, 2, 3]
         @client.allowed?(@message).must_equal true
       end
 
       it "allows all users if groups not set" do
         @message.group_ids = nil
-        @client.group_ids = [77,0,10]
+        @client.group_ids = [77, 0, 10]
         @client.allowed?(@message).must_equal true
       end
-
     end
   end
-
 end
