@@ -2,7 +2,6 @@ require_relative '../../spec_helper'
 require 'message_bus'
 
 class FakeAsync
-
   attr_accessor :cleanup_timer
 
   def <<(val)
@@ -11,18 +10,27 @@ class FakeAsync
     @sent << val
   end
 
-  def sent; @sent; end
-  def done; @done = true; end
-  def done?; @done; end
+  def sent
+    @sent
+  end
+
+  def done
+    @done = true
+  end
+
+  def done?
+    @done
+  end
 end
 
 class FakeTimer
   attr_accessor :cancelled
-  def cancel; @cancelled = true; end
+  def cancel
+    @cancelled = true
+  end
 end
 
 describe MessageBus::ConnectionManager do
-
   before do
     @bus = MessageBus
     @manager = MessageBus::ConnectionManager.new(@bus)
@@ -35,7 +43,7 @@ describe MessageBus::ConnectionManager do
   end
 
   it "should cancel the timer after its responds" do
-    m = MessageBus::Message.new(1,1,"test","data")
+    m = MessageBus::Message.new(1, 1, "test", "data")
     m.site_id = 10
     @manager.notify_clients(m)
     @client.cleanup_timer.cancelled.must_equal true
@@ -45,26 +53,22 @@ describe MessageBus::ConnectionManager do
     @manager.lookup_client(@client.client_id).must_equal @client
   end
 
-  it "should be subscribed to a channel" do
-    @manager.stats[:subscriptions][10]["test"].length == 1
-  end
-
   it "should not notify clients on incorrect site" do
-    m = MessageBus::Message.new(1,1,"test","data")
+    m = MessageBus::Message.new(1, 1, "test", "data")
     m.site_id = 9
     @manager.notify_clients(m)
     assert_nil @resp.sent
   end
 
   it "should notify clients on the correct site" do
-    m = MessageBus::Message.new(1,1,"test","data")
+    m = MessageBus::Message.new(1, 1, "test", "data")
     m.site_id = 10
     @manager.notify_clients(m)
     @resp.sent.wont_equal nil
   end
 
   it "should strip site id and user id from the payload delivered" do
-    m = MessageBus::Message.new(1,1,"test","data")
+    m = MessageBus::Message.new(1, 1, "test", "data")
     m.user_ids = [1]
     m.site_id = 10
     @manager.notify_clients(m)
@@ -74,7 +78,7 @@ describe MessageBus::ConnectionManager do
   end
 
   it "should not deliver unselected" do
-    m = MessageBus::Message.new(1,1,"test","data")
+    m = MessageBus::Message.new(1, 1, "test", "data")
     m.user_ids = [5]
     m.site_id = 10
     @manager.notify_clients(m)
@@ -83,7 +87,6 @@ describe MessageBus::ConnectionManager do
 end
 
 describe MessageBus::ConnectionManager, "notifying and subscribing concurrently" do
-
   it "does not subscribe incorrect clients" do
     manager = MessageBus::ConnectionManager.new
 
@@ -114,7 +117,7 @@ describe MessageBus::ConnectionManager, "notifying and subscribing concurrently"
 
     subscriber_threads = 10.times.map do |id|
       Thread.new do
-        m = MessageBus::Message.new(1,id,"test","data_#{id}")
+        m = MessageBus::Message.new(1, id, "test", "data_#{id}")
         m.site_id = 10
         @manager.notify_clients(m)
         1
@@ -124,5 +127,4 @@ describe MessageBus::ConnectionManager, "notifying and subscribing concurrently"
     client_threads.each(&:join).map(&:value).must_equal([1] * 10)
     subscriber_threads.each(&:join).map(&:value).must_equal([1] * 10)
   end
-
 end
